Remove cart from credit-card payment method and require button text
